pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Анатолий Зайнутдинов <zainutdinov.aaa@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
